+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New #16: Add `sendContact` method.
 - New #17: Add `getUserProfilePhotos` method and `UserProfilePhotos` type.
+- New #18: Add `sendPoll` method and `InputPollOption` type.
 
 ## 0.1.0 June 10, 2024